Use pytest to run websocket doctests
