+++ setup.cfg
@@ -11,7 +11,8 @@ norecursedirs =
     venv
     .venv
 
-collect_ignore = ['setup.py']
+rootdir = .
+collect_ignore = ['setup.py', 'src/polyswarmd/websockets/scripts/genstubs.py']
 
 [isort]
 skip_glob=tests/conftest.py
